int spec

+++ sparky/tsvDiffCalculate.py
@@ -259,7 +259,7 @@ if __name__ == "__main__":
     pd.options.display.max_columns = None
     pd.options.display.max_rows = None
 
-    dtypes={'articleid':np.int64, 'namespace':np.int64, 'anon':np.str, 'deleted':np.bool, 'revert':np.str, 'reverteds':np.str,'revid':np.int64, 'date_time':np.str, 'year':np.int8, 'month':np.int8, 'regexes':np.str, 'core_regexes':np.str,'regexes_prev':np.str, 'core_prev':np.str, 'regexes_diff_bool':np.int8, 'core_diff_bool':np.int8}
+    dtypes={'articleid':np.int64, 'namespace':np.int64, 'anon':np.str, 'deleted':np.bool, 'revert':np.str, 'reverteds':np.str,'revid':np.int64, 'date_time':np.str, 'year':np.int64, 'month':np.int64, 'regexes':np.str, 'core_regexes':np.str,'regexes_prev':np.str, 'core_prev':np.str, 'regexes_diff_bool':np.int64, 'core_diff_bool':np.int64}
     parse_dates_in = ['date_time']
 
     pd_df = pd.read_csv(file_path, sep="\t", header=0, dtype=dtypes, parse_dates=parse_dates_in)
